requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
psycopg2-binary==2.9.11
python-dotenv==1.0.0